Do dynamic trigger of quiz annotations
Local quiz annotations can be regularly reloaded from localStorage so we cannot rely on the triggering of an "enter" event on an annotation instance (which can be replaced). We rather filter on the list of annotations to check it they belong to the widget. This is less efficient but more flexible.